requirements.txt
rados==2.0.0
aiohttp==3.8.3
python-telegram-bot==20.0b0
